Remove log.Printfs.
For services (like github.com/letsencrypt/boulder/) that want to log everything
via syslog instead of stderr, and expect to receive only panics via stderr, it's
nice to be able to supress stderr logging. Some packages like
https://github.com/go-sql-driver/mysql do this by providing a way to override
the logger.

However, the logging in httpdown is minimal, and can be done by the caller
instead, so it seems better to just remove it.